requirements.txt
django==4.2.5
psycopg2==2.9.7
pillow==10.0.0
django-restframework==0.0.1